+++ RoughSets/Infovision.Datamining.RoughSet/DecisionTrees/DecisionForestDummyRough.cs
@@ -35,7 +35,7 @@ namespace Infovision.Datamining.Roughset.DecisionTrees
                 Args parms = new Args(6);
                 parms.SetParameter<DataStore>(ReductGeneratorParamHelper.TrainData, data);
                 parms.SetParameter<string>(ReductGeneratorParamHelper.FactoryKey, this.ReductGeneratorFactory);
-                parms.SetParameter<decimal>(ReductGeneratorParamHelper.Epsilon, this.Epsilon);
+                parms.SetParameter<double>(ReductGeneratorParamHelper.Epsilon, this.Epsilon);
                 parms.SetParameter<PermutationCollection>(ReductGeneratorParamHelper.PermutationCollection, localPermutationCollection);
                 parms.SetParameter<bool>(ReductGeneratorParamHelper.UseExceptionRules, false);
